+++ samples/src/08_ConstantBuffer.cpp
@@ -170,14 +170,14 @@ void init_tiny_renderer(GLFWwindow* window)
     auto vert = load_file(k_asset_dir + "constant_buffer.vs.spv");
     auto frag = load_file(k_asset_dir + "constant_buffer.ps.spv");
     tr_create_shader_program(m_renderer, 
-                             vert.size(), (uint32_t*)(vert.data()), "VSMain", 
-                             frag.size(), (uint32_t*)(frag.data()), "PSMain", &m_shader);
+                             static_cast<uint32_t>(vert.size()), (uint32_t*)(vert.data()), "VSMain", 
+                             static_cast<uint32_t>(frag.size()), (uint32_t*)(frag.data()), "PSMain", &m_shader);
 #elif defined(TINY_RENDERER_DX)
     auto hlsl = load_file(k_asset_dir + "constant_buffer.hlsl");
     tr_create_shader_program(m_renderer, 
 
-                             hlsl.size(), hlsl.data(), "VSMain", 
-                             hlsl.size(), hlsl.data(), "PSMain", &m_shader);
+                             static_cast<uint32_t>(hlsl.size()), hlsl.data(), "VSMain", 
+                             static_cast<uint32_t>(hlsl.size()), hlsl.data(), "PSMain", &m_shader);
 #endif
 
     std::vector<tr_descriptor> descriptors(1);
@@ -185,8 +185,8 @@ void init_tiny_renderer(GLFWwindow* window)
     descriptors[0].count         = 1;
     descriptors[0].binding       = 0;
     descriptors[0].shader_stages = tr_shader_stage_vert;
-    tr_create_descriptor_set(m_renderer, descriptors.size(), descriptors.data(), &m_desc_set_tri);
-    tr_create_descriptor_set(m_renderer, descriptors.size(), descriptors.data(), &m_desc_set_quad);
+    tr_create_descriptor_set(m_renderer, static_cast<uint32_t>(descriptors.size()), descriptors.data(), &m_desc_set_tri);
+    tr_create_descriptor_set(m_renderer, static_cast<uint32_t>(descriptors.size()), descriptors.data(), &m_desc_set_quad);
 
     tr_vertex_layout vertex_layout = {};
     vertex_layout.attrib_count = 1;
@@ -212,7 +212,7 @@ void init_tiny_renderer(GLFWwindow* window)
         vertexData[4*2 + 0] += -0.5f;
 
         uint64_t vertexDataSize = sizeof(float) * vertexData.size();
-        uint64_t vertexStride = sizeof(float) * 4;
+        uint32_t vertexStride = sizeof(float) * 4;
         tr_create_vertex_buffer(m_renderer, vertexDataSize, true, vertexStride, &m_tri_vertex_buffer);
         memcpy(m_tri_vertex_buffer->cpu_mapped_address, vertexData.data(), vertexDataSize);
     }
@@ -232,7 +232,7 @@ void init_tiny_renderer(GLFWwindow* window)
         vertexData[4*3 + 0] += 0.5f;
 
         uint64_t vertexDataSize = sizeof(float) * vertexData.size();
-        uint64_t vertexStride = sizeof(float) * 4;
+        uint32_t vertexStride = sizeof(float) * 4;
         tr_create_vertex_buffer(m_renderer, vertexDataSize, true, vertexStride, &m_rect_vertex_buffer);
         memcpy(m_rect_vertex_buffer->cpu_mapped_address, vertexData.data(), vertexDataSize);
         
@@ -291,7 +291,7 @@ void draw_frame()
 
     tr_begin_cmd(cmd);
     tr_cmd_render_target_transition(cmd, render_target, tr_texture_usage_present, tr_texture_usage_color_attachment); 
-    tr_cmd_set_viewport(cmd, 0, 0, s_window_width, s_window_height, 0.0f, 1.0f);
+    tr_cmd_set_viewport(cmd, 0, 0, static_cast<float>(s_window_width), static_cast<float>(s_window_height), 0.0f, 1.0f);
     tr_cmd_set_scissor(cmd, 0, 0, s_window_width, s_window_height);
     tr_cmd_begin_render(cmd, render_target);
     tr_clear_value clear_value = {0.0f, 0.0f, 0.0f, 0.0f};
