+++ tinyvk.h
@@ -3731,7 +3731,7 @@ void tr_internal_vk_create_buffer(tr_renderer* p_renderer, tr_buffer* p_buffer)
     // Align the buffer size to multiples of the dynamic uniform buffer minimum size
     if (p_buffer->usage & tr_buffer_usage_uniform_cbv) {
         // Make minimum size 256 bytes to match D3D12
-        p_buffer->size = tr_round_up(tr_max(p_buffer->size, 256), 
+        p_buffer->size = tr_round_up(tr_max(static_cast<uint32_t>(p_buffer->size), 256), 
                                      static_cast<uint32_t>(p_renderer->vk_active_gpu_properties.limits.minUniformBufferOffsetAlignment));
     }
